Move scanning logic to wallet
