+++ rust/src/api/chain.rs
@@ -4,6 +4,7 @@ use crate::blindbit;
 
 pub async fn get_chain_height(blindbit_url: String) -> anyhow::Result<u32> {
     let url = Url::parse(&blindbit_url)?;
+    let blindbit_client = blindbit::BlindbitClient::new(url);
 
-    blindbit::logic::get_chain_height(url).await
+    blindbit_client.block_height().await
 }
