+++ rust/src/wallet/mod.rs
@@ -1,2 +1,3 @@
 pub mod recorded;
 pub mod spwallet;
+pub mod scan;
